% Generated by roxygen2 (4.0.2): do not edit by hand
\name{avg}
\alias{avg}
\title{Averaging functoin}
\usage{
avg(p, c, rm = 0)
}
\arguments{
\item{p}{raw pressure signal}

\item{c}{cycle detector ouput (borders)}

\item{rm}{outlier removal threshold}
}
\value{
value
}
\description{
Averaging functoin
}

