% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cal}
\alias{cal}
\title{BP calibration function}
\usage{
cal(inp, max, mean, min, central = TRUE)
}
\arguments{
\item{inp}{data file}

\item{max}{max}

\item{mean}{mean}

\item{min}{min}

\item{central}{central or peripheral}
}
\value{
value
}
\description{
BP calibration function
}

