% Generated by roxygen2 (4.0.2): do not edit by hand
\name{foot}
\alias{foot}
\title{Intersecting tangents function}
\usage{
foot(bp, c, d)
}
\arguments{
\item{bp}{raw blood pressure recording}

\item{c}{ECG-based cycle delimiter}

\item{d}{draw}
}
\value{
value
}
\description{
Intersecting tangents function
}

