% Generated by roxygen2 (4.0.2): do not edit by hand
\name{out}
\alias{out}
\title{Outlier detection function}
\usage{
out(x, z = 0.7)
}
\arguments{
\item{x}{numeric vector}

\item{z}{standardized threshold}
}
\value{
vector index of outliers
}
\description{
Outlier detection function
}
\examples{
out(runif(10), 0.9)
}

