% Generated by roxygen2 (4.0.2): do not edit by hand
\name{peak}
\alias{peak}
\title{Peak detector functoin}
\usage{
peak(p, ecg, sr = 1000, t = 0.7, dt = TRUE)
}
\arguments{
\item{p}{pressure signal}

\item{ecg}{ecg signal}

\item{sr}{sampling rate (Hz)}

\item{t}{threshold}

\item{dt}{detrend}
}
\value{
value
}
\description{
Peak detector functoin
}

