% Generated by roxygen2 (4.0.2): do not edit by hand
\name{proc}
\alias{proc}
\title{Process function}
\usage{
proc(adat, ekg = TRUE, thres = 0.9, freq = 1000, dt = TRUE)
}
\arguments{
\item{adat}{raw, unprocessed data file}

\item{ekg}{pressure/ECG-based cycle recognition}

\item{thres}{peak detector threshold}

\item{freq}{smapling rate}

\item{dt}{detrending}
}
\value{
value
}
\description{
Process function
}

