% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wfp}
\alias{wfp}
\title{Wave-free period detector function}
\usage{
wfp(p, inc, end)
}
\arguments{
\item{p}{averaged, scaled pressure wave}

\item{inc}{location of incisura (end-systole)}

\item{end}{end of the pressure wave}
}
\value{
value
}
\description{
Wave-free period detector function
}

